fix word count, move function to alt_analyze.py, filter out words with length < 3
