+++ tests/test_alt_analyze.py
@@ -42,7 +42,7 @@ def test_analyze_linguistic_complexity():
     text="The implementation of sophisticated algorithms necessitates comprehensive understanding of computational complexity theory. Contemporary methodologies incorporate various optimization techniques to enhance performance metrics."
     metrics=analyze_linguistic_complexity(text)
 
-    assert metrics ['word_count']==24
+    assert metrics ['word_count']==22
     assert metrics ['sentence_count']==2
     assert metrics['char_count']==226
     assert metrics['flesch_kincaid_grade']>10
